+++ src/client/dcp_client/utils/settings.py
@@ -1,4 +1,5 @@
-def init():
+def init() -> None:
+    """ Initialise global variables."""
     global accepted_types
     accepted_types = (".jpg", ".jpeg", ".png", ".tiff", ".tif")
     global seg_name_string
